api token

+++ Laravel/projecte_final/routes/api.php
@@ -9,7 +9,7 @@ use App\Http\Controllers\ServeiController;
 use App\Http\Controllers\Controller;
 use App\Http\Controllers\TreballadorController;
 use App\Http\Controllers\UserController;
-
+use App\Http\Controllers\LoginController;
 
 
 
@@ -33,89 +33,87 @@ use App\Http\Controllers\ApiController;
 Route::get('/', [Controller::class, 'home'])->name('home');
 
 //// Cites
+Route::get('/citas', [CitaController::class, 'getCitas'])->middleware('auth');
 
-Route::get('/citas', [CitaController::class, 'getCitas']);
+Route::get('/cita/{id}', [CitaController::class, 'getCita'])->middleware('auth');
 
-Route::get('/cita/{id}', [CitaController::class, 'getCita']);
+Route::post('/cita/{id}', [CitaController::class, 'updateCita'])->name('cita_update')->middleware('auth');
 
-Route::post('/cita/{id}', [CitaController::class, 'updateCita'])->name('cita_update');
+Route::post('/cita', [CitaController::class, 'insertCita'])->middleware('auth');
 
-Route::post('/cita', [CitaController::class, 'insertCita']);
-
-Route::delete('/cita/{id}', [CitaController::class, 'deleteCita'])->name('cita_delete');
+Route::delete('/cita/{id}', [CitaController::class, 'deleteCita'])->name('cita_delete')->middleware('auth');
 
 //// Clients
 
-Route::get('/clients', [ClientController::class, 'getClients'])->name('clients_list');
+Route::get('/clients', [ClientController::class, 'getClients'])->name('clients_list')->middleware('auth');
 
-Route::get('/client/{id}', [ClientController::class, 'getClient'])->name('client_cerca');
+Route::get('/client/{id}', [ClientController::class, 'getClient'])->name('client_cerca')->middleware('auth');
 
-Route::post('/client/{id}', [ClientController::class, 'updateClient'])->name('client_update');
+Route::post('/client/{id}', [ClientController::class, 'updateClient'])->name('client_update')->middleware('auth');
 
-Route::post('/client', [ClientController::class, 'insertClient']);
+Route::post('/client', [ClientController::class, 'insertClient'])->middleware('auth');
 
-Route::delete('/client/{id}', [ClientController::class, 'deleteClient'])->name('client_delete');
+Route::delete('/client/{id}', [ClientController::class, 'deleteClient'])->name('client_delete')->middleware('auth');
 
-Route::get('/client-user/{id}', [ClientController::class, 'getClientByUserId'])->name('client_user');
-
+Route::get('/client-user/{id}', [ClientController::class, 'getClientByUserId'])->name('client_user')->middleware('auth');
 
 //// Treballadors
 
-Route::get('/treballadors', [TreballadorController::class, 'getTreballadors'])->name('treballadors_list');
+Route::get('/treballadors', [TreballadorController::class, 'getTreballadors'])->name('treballadors_list')->middleware('auth');
 
-Route::get('/treballador/{id}', [TreballadorController::class, 'getTreballador'])->name('treballador_cerca');
+Route::get('/treballador/{id}', [TreballadorController::class, 'getTreballador'])->name('treballador_cerca')->middleware('auth');
 
-Route::post('/treballador/{id}', [TreballadorController::class, 'updateTreballador'])->name('treballador_update');
+Route::post('/treballador/{id}', [TreballadorController::class, 'updateTreballador'])->name('treballador_update')->middleware('auth');
 
-Route::post('/treballador', [TreballadorController::class, 'insertTreballador'])->name('treballador_insert');
+Route::post('/treballador', [TreballadorController::class, 'insertTreballador'])->name('treballador_insert')->middleware('auth');
 
-Route::delete('/treballador/{id}', [TreballadorController::class, 'deleteTreballador'])->name('treballador_delete');
+Route::delete('/treballador/{id}', [TreballadorController::class, 'deleteTreballador'])->name('treballador_delete')->middleware('auth');
 
-Route::get('/treballador-user/{id}', [TreballadorController::class, 'getTreballadorByUserId'])->name('treballador_user');
-
+Route::get('/treballador-user/{id}', [TreballadorController::class, 'getTreballadorByUserId'])->name('treballador_user')->middleware('auth');
 
 //// Users
 
-Route::get('/users', [UserController::class, 'getUsers'])->name('users_list');
+Route::get('/users', [UserController::class, 'getUsers'])->name('users_list')->middleware('auth');
 
-Route::get('/user/{id}', [UserController::class, 'getUser'])->name('user_cerca');
+Route::get('/user/{id}', [UserController::class, 'getUser'])->name('user_cerca')->middleware('auth');
 
+// Login
+
+Route::post('/login', [LoginController::class, 'login'])->name('user_login');
 
 // Buscar usuario por correo para login
 Route::get('/userLogin/{email}', [UserController::class, 'getUserLogin'])->name('user_login');
 
+Route::post('/user/{id}', [UserController::class, 'updateUser'])->name('user_update')->middleware('auth');
 
-Route::post('/user/{id}', [UserController::class, 'updateUser'])->name('user_update');
+Route::post('/user', [UserController::class, 'insertUser'])->middleware('auth');
 
-Route::post('/user', [UserController::class, 'insertUser']);
-
-Route::delete('/user/{id}', [UserController::class, 'deleteUser'])->name('user_delete');
+Route::delete('/user/{id}', [UserController::class, 'deleteUser'])->name('user_delete')->middleware('auth');
 
 //// Vehicles
 
-Route::get('/vehicles', [VehicleController::class, 'getVehicles'])->name('vehicles_list');
+Route::get('/vehicles', [VehicleController::class, 'getVehicles'])->name('vehicles_list')->middleware('auth');
 
-Route::get('/vehicle/{id}', [VehicleController::class, 'getVehicle'])->name('vehicle_cerca');
+Route::get('/vehicle/{id}', [VehicleController::class, 'getVehicle'])->name('vehicle_cerca')->middleware('auth');
 
-Route::get('/vehicle-client/{idClient}', [VehicleController::class, 'getVehicleClient'])->name('vehicle_client');
+Route::get('/vehicle-client/{idClient}', [VehicleController::class, 'getVehicleClient'])->name('vehicle_client')->middleware('auth');
 
-Route::post('/vehicle/{id}', [VehicleController::class, 'updateVehicle'])->name('vehicle_update');
+Route::post('/vehicle/{id}', [VehicleController::class, 'updateVehicle'])->name('vehicle_update')->middleware('auth');
 
-Route::post('/vehicle', [VehicleController::class, 'insertVehicle'])->name('vehicle_insert');
+Route::post('/vehicle', [VehicleController::class, 'insertVehicle'])->name('vehicle_insert')->middleware('auth');
 
-Route::delete('/vehicle/{id}', [VehicleController::class, 'deleteVehicle'])->name('vehicle_delete');
+Route::delete('/vehicle/{id}', [VehicleController::class, 'deleteVehicle'])->name('vehicle_delete')->middleware('auth');
 
 //// Serveis
 
-Route::get('/serveis', [ServeiController::class, 'getServeis'])->name('serveis_list');
+Route::get('/serveis', [ServeiController::class, 'getServeis'])->name('serveis_list')->middleware('auth');
 
-Route::get('/servei/{id}', [ServeiController::class, 'getServei'])->name('servei_cerca');
+Route::get('/servei/{id}', [ServeiController::class, 'getServei'])->name('servei_cerca')->middleware('auth');
 
-Route::post('/servei/{id}', [ServeiController::class, 'updateServei'])->name('servei_update');
+Route::post('/servei/{id}', [ServeiController::class, 'updateServei'])->name('servei_update')->middleware('auth');
 
-Route::post('/servei', [ServeiController::class, 'insertServei'])->name('servei_insert');
+Route::post('/servei', [ServeiController::class, 'insertServei'])->name('servei_insert')->middleware('auth');
 
-Route::delete('/servei/{id}', [ServeiController::class, 'deleteServei'])->name('servei_delete');
-
+Route::delete('/servei/{id}', [ServeiController::class, 'deleteServei'])->name('servei_delete')->middleware('auth');
